Workflow structure organise for functions 'check_for_file' , 'has_ocr', 'extract_text_doctr', 'query_prompt', 'get_category_type'
