+++ src/processors/document_tagger/__init__.py
@@ -1,0 +1,3 @@
+from .document_category import get_category_type
+
+__all__ = ['get_category_type']
